Update singly_linked_list.py
This file contains the implementation of a singly linked list. It consists of two classes: `Node` and `LinkedList`.

The `Node` class represents a single node in the linked list. Each node contains a value and a reference to the next node in the list. The `__repr__` method is overridden to provide a string representation of the node's value and the ID of the next node.

The `LinkedList` class represents the linked list itself. It maintains a reference to the head node of the list. The list supports various operations such as adding elements to the front (`push_front`), inserting elements at a specified position (`insert`), adding elements to the back (`push_back`), finding the position of a specific value (`find`), deleting an element at a specified position (`delete`), and obtaining the length of the list (`__len__`). It also implements an iterator (`__iter__`) to allow iterating over the elements of the list. The `__repr__` method is overridden to provide a string representation of the list by displaying the values of all the nodes.

Overall, this file provides a basic implementation of a singly linked list data structure with common operations.