Create general_tree.py
Implement a generalized tree (general tree), whose nodes store functions as values. Allow to add or remove a function contained in any node. Each function has the same signature: it receives an integer, it returns an integer. It is necessary to implement the function execute(int) for the tree, which receives a number and calls the functions contained in all nodes of the tree in turn, passing the received number to the function, and passing the return value of the function to the function of the next node as an argument.